+++ tests/unit/serviceIntegrity.js
@@ -82,10 +82,15 @@ describe('serviceIntegrity', function() {
 
             it('should return resolved promise', function() {
 
-                return serviceIntegrity.inspect(this.app).should.become({
+                return serviceIntegrity.inspect(this.app, {
+                    inspectors: {
+                        services: serviceIntegrity.inspectDependentServices(this.app)
+                    }
+                }).should.become({
                     node          : true,
                     postgres      : true,
                     couchbase     : true,
+                    services      : true,
                     configuration : true,
                     session       : true
                 });